more refactoring __init__
utility readableList() to produce list of options from dict_key structure
move grouptree and ownertree to profile_publishers.py
